tracker/management/commands/handlers.py
# handlers.py
# Telegram command and callback handlers
from telegram.ext import CommandHandler, CallbackQueryHandler

# Import your handler functions here, e.g.
# from .run_bot import start, help_command, ...

# Example registration function
def register_handlers(dispatcher, handler_funcs):
    dispatcher.add_handler(CommandHandler("start", handler_funcs["start"]))
    dispatcher.add_handler(CommandHandler("help", handler_funcs["help_command"]))
    # Add more handlers as needed
    # dispatcher.add_handler(CallbackQueryHandler(...))
