requirements.txt
Django>=5.1,<7
djangorestframework>=3.15
python-telegram-bot>=20.7
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
Pillow>=10.4.0
requests>=2.31.0
gunicorn>=23.0.0
whitenoise>=6.7.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
aiohttp>=3.9.0
emoji>=2.12.1
